Re-routed the mappers to DAO interfaces

+++ src/main/java/com/solvd/developmentCompany/interfaces/IBaseDAO.java
@@ -3,9 +3,9 @@ package com.solvd.developmentCompany.interfaces;
 import java.util.List;
 
 public interface IBaseDAO<T> {
-    T getById(int id);
+    T getById(Long id);
     List<T> getAll();
     void save(T entity);
     void update(T entity);
-    void delete(int id);
+    void delete(Long id);
 }
